+++ CHANGES.md
@@ -2,6 +2,31 @@
 
 All notable changes to the webmate Java SDK will be documented in this file.
 
+## [0.23] - 2018-12-04
+###Fixes
+- Fixes rare Serialization Issue
+
+
+## [0.22] - 2018-11-27
+###New
+- Added Selenium Service API methods
+- Added Browser Utils to retrieve the most recent Browsers available in a given Project
+
+## [0.21] - 2018-10-05
+
+### Fixes
+- Added timeout to BrowserSessionRefs createState Method
+
+## [0.20] - 2018-09-05
+
+### Fixes
+- Racecondition fixed when creating jobs
+
+## [0.19] - 2018-08-08
+
+### Fixes
+- Ensures compatibility with current Webmate Release. NOTE: prior versions will no longer work!
+
 ## [0.18] - 2018-06-12
 
 ### New